+++ credit/src/test/java/org/cryptimeleon/incentivesystem/services/credit/TestConfiguration.java
@@ -1,6 +1,6 @@
 package org.cryptimeleon.incentivesystem.services.credit;
 
-import org.cryptimeleon.incentivesystem.cryptoprotocol.interfaces.provider.CreditInterface;
+import org.cryptimeleon.incentivesystem.services.credit.interfaces.CreditInterface;
 import org.cryptimeleon.incentivesystem.services.credit.mock.TestBasketServerClientMock;
 import org.cryptimeleon.incentivesystem.services.credit.mock.TestCryptoCreditMock;
 import org.cryptimeleon.incentivesystem.services.credit.model.interfaces.BasketServerClientInterface;