+++ cryptoprotocol/src/main/java/org/cryptimeleon/incentivesystem/cryptoprotocol/exceptions/SPSEQException.java
@@ -1,7 +1,0 @@
-package org.cryptimeleon.incentivesystem.cryptoprotocol.exceptions;
-
-public class SPSEQException extends Exception {
-    public SPSEQException(String message) {
-        super(message);
-    }
-}