+++ cryptoprotocol/src/main/java/org/cryptimeleon/incentivesystem/cryptoprotocol/interfaces/provider/CreditInterface.java
@@ -1,8 +1,0 @@
-package org.cryptimeleon.incentivesystem.cryptoprotocol.interfaces.provider;
-
-public interface CreditInterface {
-    // pp, pk, sk are provided via constructor of the class implementing this interface
-
-    // earnAmount (k) can be assumed to be verified by credit service (using basket server)
-    String computeSerializedResponse(String serializedEarnRequest, long earnAmount);
-}