+++ cryptoprotocol/src/main/java/org/cryptimeleon/incentivesystem/cryptoprotocol/interfaces/provider/IssueInterface.java
@@ -1,8 +1,0 @@
-package org.cryptimeleon.incentivesystem.cryptoprotocol.interfaces.provider;
-
-public interface IssueInterface {
-    // Provider is stateless (besides pp, pk, sk) => provide all information of request via parameters
-    // pp, pk, sk provided via constructor of the class implementing this interface
-
-    String computeSerializedIssueReponse(String serializedJoinRequest, String serializedUserPublicKey);
-}