+++ cryptoprotocol/src/main/java/org/cryptimeleon/incentivesystem/cryptoprotocol/model/messages/EarnRequestMessage.java
@@ -1,9 +1,0 @@
-package org.cryptimeleon.incentivesystem.cryptoprotocol.model.messages;
-
-import org.cryptimeleon.craco.sig.sps.eq.SPSEQSignature;
-import org.cryptimeleon.math.structures.groups.GroupElement;
-
-public class EarnRequestMessage {
-    private GroupElement blindedCommitment;
-    private SPSEQSignature blindedCertificate;
-}